+++ src/pages/Home.jsx
@@ -1,53 +1,18 @@
 import React from "react";
-import { useNavigate } from "react-router-dom";
 import Header from "../components/Header";
 
 export default function Home() {
-  const navigate = useNavigate();
-  const token = localStorage.getItem("token");
-
   return (
-    <div style={{ display: "flex", justifyContent: "center", gap: "30px" }}>
-      
+    <div
+      style={{
+        height: "1500px",
+        backgroundColor: "gray",
+        display: "flex",
+        flexDirection: "column",
+      }}
+    >
       <Header />
-      
-      Trang chủ
-      {token === null ? (
-        <div>
-          <button
-            onClick={() => {
-              navigate("/register");
-            }}
-          >
-            Đăng ký
-          </button>
-          <button
-            onClick={() => {
-              navigate("/login");
-            }}
-          >
-            Đăng nhập
-          </button>
-        </div>
-      ) : (
-        <div>
-          <button
-            onClick={() => {
-              navigate("/profile");
-            }}
-          >
-            Tài khoản của bạn
-          </button>
-          <button
-            onClick={() => {
-              localStorage.removeItem("token");
-              navigate("/");
-            }}
-          >
-            Đăng xuất
-          </button>
-        </div>
-      )}
+      <span>Trang chủ</span>
     </div>
   );
 }
