+++ src/styles/Header.scss
@@ -1,0 +1,147 @@
+.header {
+    position: fixed;
+    display: flex;
+    align-items: center;
+    width: 100%;
+    height: 100px;
+    background-color: #ee4d2d;
+    z-index: 100;
+
+    &__logo {
+        margin-left: 8%;
+        width: 150px;
+        & img {
+            width: 100%;
+            cursor: pointer;
+        }
+    }
+
+    &__search {
+        position: absolute;
+        top: 50%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+        width: 50%;
+        height: 40%;
+
+        & input {
+            box-sizing: border-box;
+            margin: 0;
+            padding-left: 10px;
+            padding-right: 80px;
+            width: 100%;
+            height: 100%;
+            border: 0;
+            border-radius: 4px;
+            font-size: 16px;
+        }
+
+        & .btnSearch {
+            position: absolute;
+            top: 50%;
+            right: 1%;
+            transform: translateY(-50%);
+            width: 8%;
+            height: 80%;
+            background-color: #ee4d2d;
+            border: 0;
+            border-radius: 4px;
+            color: #fff;
+            cursor: pointer;
+            font-size: 18px;
+
+            & svg {
+                position: absolute;
+                top: 50%;
+                left: 50%;
+                transform: translate(-50%, -50%);
+            }
+        }
+    }
+
+    &__user {
+        position: absolute;
+        top: 10px;
+        right: 0;
+        transform: translate(-50%, -50%);
+
+        & button {
+            background-color: transparent;
+            border: 0;
+            color: #fff;
+            cursor: pointer;
+        }
+    }
+}
+
+/////////////////////////////////////////////////////////////////Tablet
+@media only screen and (min-width: 740px) and (max-width: 1023px) {
+    .header {
+        &__logo {
+            margin-left: 8%;
+        }
+    
+        &__search {
+            position: absolute;
+            top: 50%;
+            left: 60%;
+            transform: translate(-50%, -50%);
+            width: 40%;
+            height: 40%;
+
+            & input {
+                padding-left: 10px;
+                padding-right: 50px;
+            }
+        }
+    
+        &__user {
+            position: absolute;
+            top: 10px;
+            right: 0;
+            transform: translate(-30%, -50%);
+        }
+    }
+}
+
+/////////////////////////////////////////////////////////////Mobile
+@media only screen and (max-width: 739px) {
+    .header {
+        &__logo {
+            margin-left: 5%;
+        }
+    
+        &__search {
+            position: absolute;
+            top: 50%;
+            left: 70%;
+            transform: translate(-50%, -50%);
+            width: 40%;
+            height: 40%;
+
+            & input {
+                padding-left: 10px;
+                padding-right: 50px;
+            }
+
+            & .btnSearch {        
+                width: 15%;  
+                right: 2%;             
+    
+                & svg {
+                    position: absolute;
+                    top: 50%;
+                    left: 50%;
+                    transform: translate(-50%, -50%);
+                }
+            }
+        }
+    
+        &__user {
+            position: absolute;
+            top: 10px;
+            right: 0;
+            transform: translate(-10%, -50%);
+        }
+    }
+}
