+++ src/styles/Profile.scss
@@ -1,6 +1,5 @@
-
-
 .background {
+    padding-top: 100px;
     width: 100%;
     height: 100vh;
     background-color: gray;
