+++ src/styles/Search.scss
@@ -1,4 +1,5 @@
 .products {
+    padding-top: 100px;
     width: 60%;
     margin: 0 auto;
     display: flex;
@@ -7,16 +8,14 @@
     justify-content: center;
 }
 
-
-
 .product {
     width: 186px;
     height: auto;
     font-size: 0.75rem;
-    color: rgba(0,0,0,.87);
+    color: rgba(0, 0, 0, 0.87);
     background-color: white;
     border: 1px solid transparent;
-    box-shadow: 0 0.0625rem 0.125rem 0 rgba(0, 0, 0, 10%); 
+    box-shadow: 0 0.0625rem 0.125rem 0 rgba(0, 0, 0, 10%);
     border-radius: 0.125rem;
 }
 
@@ -31,8 +30,6 @@
     height: 100%;
     object-fit: cover;
 }
-
-
 
 .product__name {
     padding: 8px;
